requirements.txt
opensearch-py
sentence-transformers
pypdf
transformers
torch
python-dotenv
accelerate
google-generativeai

